:art: 优化项目格式
