--- src/alert/index.css
@@ -1,3 +1,0 @@
-.common{
-    margin-bottom: 20px;
-}
