--- src/button/index.css
@@ -1,3 +1,0 @@
-.common{
-    margin-right: 20px;
-}
