--- src/icon/index.css
@@ -1,6 +1,0 @@
-.common{
-    margin-bottom: 20px;
-}
-.common span{
-    margin-right: 10px;
-}
